third_party/blink/web_tests/external/wpt/event-timing/first-input-interactionid-tap.html

<!DOCTYPE html>
<html>
<meta charset=utf-8 />
<meta name="timeout" content="long">
<title>Event Timing: first-input-interactionId-tap.</title>
<button id='test'>Tap</button>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/testdriver.js></script>
<script src=/resources/testdriver-actions.js></script>
<script src=/resources/testdriver-vendor.js></script>
<script src=resources/event-timing-test-utils.js></script>

<script>
  let firstInputInteractionId = 0;
  let eventTimingPointerDownInteractionId = 0;
  let hasFirstInputEntry = false;
  let hasEventTimingPointerDownEntry = false;

  const isReadyForCheck = () => {
    return hasFirstInputEntry && hasEventTimingPointerDownEntry;
  }

  async_test(function (t) {
    assert_implements(window.PerformanceEventTiming, 'Event Timing is not supported.');

    new PerformanceObserver(t.step_func(entryList => {
      entryList.getEntries().forEach(entry => {
        switch (entry.entryType) {
          case "first-input": {
            firstInputInteractionId = entry.interactionId;
            hasFirstInputEntry = true;
            break;
          }
          case "event": {
            if ('pointerdown' == entry.name) {
              eventTimingPointerDownInteractionId = entry.interactionId;
              hasEventTimingPointerDownEntry = true;
            }
            break;
          }
        }
      })

      if (isReadyForCheck()) {
        assert_greater_than(firstInputInteractionId, 0, 'The first input entry should have a non-trivial interactionId');
        assert_equals(firstInputInteractionId, eventTimingPointerDownInteractionId, 'The first input entry should have the same interactionId as the event timing pointerdown entry');
        t.done();
      }
    })).observe({ entryTypes: ["event", "first-input"] });

    addListenersAndTap(document.getElementById('test'), ['pointerdown', 'pointerup']);
  }, "Event Timing: The interactionId of first input should match the event timing pointerdown entry when tap.");
</script>

</html>
